+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Release History
 
+- Version [2.2.0](#v2.2.0)
 - Version [2.1.0](#v2.1.0)
 - Version [2.0.1](#v2.0.1)
 - Version [2.0.0](#v2.0.0)
@@ -17,6 +18,15 @@
 - Version [0.6.0](#v0.6.0)
 - Version [0.5.2](#v0.5.2)
 - Version [0.5.1](#v0.5.1)
+
+## <a name="v2.2.0">Release 2.2.0</a>
+
+### Changed
+
+- Remove default department from organizations
+- Remove organization id
+- workaround for function(address,bytes32) signature breaking with empty bytes32:
+  `AddressScopes.resolveAddressScope` returns a marker for an empty value rather than "".
 
 ## <a name="v2.1.0">Release 2.1.0</a>